Force UseOROCOS.cmake to check if all packages metioned in package.xml are present in system.
This change solves ambigous errors when typekit A depends on typekit B
but typekit B does not present in system. Then build system compile
typekit A without error but attempt to compile any package depent
on typekit A will causes error.

Th reason is simple. .pc file of typekit A contains reference on
typekit B which does not present in system.